docs(input-field): use a more pedagogical example
The example that showed two fields in a row wasn't so useful I believe.
So I removed it and added an example that shows an important
layout issue, if people just stack fields on top of each other.
