--- src/components/input-field/examples/input-field-text-inline.scss
@@ -1,12 +1,0 @@
-section {
-    position: relative;
-
-    limel-input-field {
-        display: inline-block;
-        width: calc(50% - 0.625rem);
-
-        &:first-of-type {
-            margin-right: 1.25rem;
-        }
-    }
-}
