+++ src/components/input-field/examples/input-field-text-multiple.scss
@@ -1,0 +1,9 @@
+section {
+    display: grid;
+}
+
+.add-distance {
+    section {
+        grid-gap: 1rem;
+    }
+}
